backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
dj-database-url==2.1.0
Django==5.2.6
django-cors-headers==4.8.0
django-filter==24.3
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.8.2
